[neovim] cleanup lazy config, fixed cland lsp
cleanup lsp configuration
move togglebg configuraion to a dedicated file
move gitsigns-setup to a dedicated file
