+++ neovim/init.lua
@@ -33,6 +33,9 @@ require 'lsp-setup'
 require 'cmp-setup'
 
 require 'nvimtree-setup'
+
+require 'gitsigns-setup'
+
 -- The line beneath this is called `modeline`. See `:help modeline`
 -- vim: ts=2 sts=2 sw=2 et
