+++ neovim/lua/lazy-plugins.lua
@@ -1,19 +1,3 @@
--- initialize the colorscheme for the first run
--- hacky way of detecting the current theme
--- todo-find a DE independent way of getting this
--- see https://arslan.io/2021/02/15/automatic-dark-mode-for-terminal-applications/
-local function redrawLine()
-  require('lualine').setup()
-end
-
-local function setBackground()
-  if string.match(vim.fn.system({ 'gsettings', 'get', 'org.gnome.desktop.interface', 'color-scheme' }), '.*dark') then
-    vim.o.background = 'dark'
-  else
-    vim.o.background = 'light'
-  end
-end
-
 -- [[ Configure plugins ]]
 -- NOTE: Here is where you install your plugins.
 --  You can configure plugins using the `config` key.
@@ -25,6 +9,8 @@ require('lazy').setup({
   -- TODO: try this plugin 
   -- use {'ojroques/nvim-osc52'}
 
+  -- collow any #<color> to its value
+  'ap/vim-css-color',
   -- Detect tabstop and shiftwidth automatically
   'tpope/vim-sleuth',
   --  [tmux]
@@ -32,7 +18,8 @@ require('lazy').setup({
   'christoomey/vim-tmux-navigator',
   'christoomey/vim-tmux-runner',
   'tmux-plugins/vim-tmux',
-
+    -- Adds git related signs to the gutter, as well as utilities for managing changes
+  'lewis6991/gitsigns.nvim',
   -- NOTE: This is where your plugins related to LSP can be installed.
   --  The configuration is done below. Search for lspconfig to find it below.
   {
@@ -50,30 +37,13 @@ require('lazy').setup({
       'folke/neodev.nvim',
     },
   },
-  -- TODO prepare the dev env for C++
-  -- check if clangd_extensions can be configured to work with mason
-
-  -- use 'hrsh7th/cmp-nvim-lsp'
-  -- use 'hrsh7th/cmp-buffer'
-  -- use 'hrsh7th/cmp-path'
-  -- use 'hrsh7th/cmp-cmdline'
-  -- use 'hrsh7th/nvim-cmp'
-  -- use 'hrsh7th/vim-vsnip'
-  -- use 'hrsh7th/vim-vsnip-integ'
   -- Configurations for Nvim LSP
-  -- use 'neovim/nvim-lspconfig'
-  -- use {
-  --       'nvim-treesitter/nvim-treesitter',
-  --       run = function()
-  --           local ts_update = require('nvim-treesitter.install').update({ with_sync = true })
-  --           ts_update()
-  --       end,
-  -- }
-  -- 'p00f/clangd_extensions.nvim',
+  'p00f/clangd_extensions.nvim',
+  -- TODO probably this is not needed and most likely covered already by mason
   -- { 'rhysd/vim-clang-format',
   --     build = function()
   --       vim.g['clang_format#detect_style_file'] = 1
-  --     end
+  --     end,
   -- },
 
   {
@@ -83,7 +53,9 @@ require('lazy').setup({
       -- Snippet Engine & its associated nvim-cmp source
       'L3MON4D3/LuaSnip',
       'saadparwaiz1/cmp_luasnip',
-
+      -- cmdline autocompletion
+      'hrsh7th/cmp-buffer',
+      'hrsh7th/cmp-cmdline',
       -- Adds LSP completion capabilities
       'hrsh7th/cmp-nvim-lsp',
       'hrsh7th/cmp-path',
@@ -99,8 +71,6 @@ require('lazy').setup({
       'nvim-tree/nvim-web-devicons' -- optional, for file icons
     },
   },
-  -- colors
-  'ap/vim-css-color',
   -- Useful plugin to show you pending keybinds.
   { 'folke/which-key.nvim',  opts = {} },
    -- Git related plugins
@@ -114,86 +84,10 @@ require('lazy').setup({
       --this is to include selfhosed bitbucket instances that are not worth 
       --shareing in a public configuration. if the file will not be present 
       --we'll just ignore it
-      local status, lfs = pcall(require, "fubitive_cfg")
-      if (status) then
-        print(lfs)   --lfs exists, so use it.
-      end
+      pcall(require, "fubitive_cfg")
     end
   },
-  {
-    -- Adds git related signs to the gutter, as well as utilities for managing changes
-    'lewis6991/gitsigns.nvim',
-    opts = {
-      -- See `:help gitsigns.txt`
-      signs = {
-        add = { text = '+' },
 
-        delete = { text = '_' },
-        topdelete = { text = '‾' },
-        changedelete = { text = '~' },
-      },
-      on_attach = function(bufnr)
-        local gs = package.loaded.gitsigns
-
-        local function map(mode, l, r, opts)
-          opts = opts or {}
-          opts.buffer = bufnr
-          vim.keymap.set(mode, l, r, opts)
-        end
-
-        -- Navigation
-        map({ 'n', 'v' }, ']c', function()
-          if vim.wo.diff then
-            return ']c'
-          end
-          vim.schedule(function()
-            gs.next_hunk()
-          end)
-          return '<Ignore>'
-        end, { expr = true, desc = 'Jump to next hunk' })
-
-        map({ 'n', 'v' }, '[c', function()
-          if vim.wo.diff then
-            return '[c'
-          end
-          vim.schedule(function()
-            gs.prev_hunk()
-          end)
-          return '<Ignore>'
-        end, { expr = true, desc = 'Jump to previous hunk' })
-
-        -- Actions
-        -- visual mode
-        map('v', '<leader>hs', function()
-          gs.stage_hunk { vim.fn.line '.', vim.fn.line 'v' }
-        end, { desc = 'stage git hunk' })
-        map('v', '<leader>hr', function()
-          gs.reset_hunk { vim.fn.line '.', vim.fn.line 'v' }
-        end, { desc = 'reset git hunk' })
-        -- normal mode
-        map('n', '<leader>hs', gs.stage_hunk, { desc = 'git stage hunk' })
-        map('n', '<leader>hr', gs.reset_hunk, { desc = 'git reset hunk' })
-        map('n', '<leader>hS', gs.stage_buffer, { desc = 'git Stage buffer' })
-        map('n', '<leader>hu', gs.undo_stage_hunk, { desc = 'undo stage hunk' })
-        map('n', '<leader>hR', gs.reset_buffer, { desc = 'git Reset buffer' })
-        map('n', '<leader>hp', gs.preview_hunk, { desc = 'preview git hunk' })
-        map('n', '<leader>hb', function()
-          gs.blame_line { full = false }
-        end, { desc = 'git blame line' })
-        map('n', '<leader>hd', gs.diffthis, { desc = 'git diff against index' })
-        map('n', '<leader>hD', function()
-          gs.diffthis '~'
-        end, { desc = 'git diff against last commit' })
-
-        -- Toggles
-        map('n', '<leader>tb', gs.toggle_current_line_blame, { desc = 'toggle git blame line' })
-        map('n', '<leader>td', gs.toggle_deleted, { desc = 'toggle git show deleted' })
-
-        -- Text object
-        map({ 'o', 'x' }, 'ih', ':<C-U>Gitsigns select_hunk<CR>', { desc = 'select git hunk' })
-      end,
-    },
-  },
   -- [themes]
   {
     'mikker/vim-togglebg',
@@ -203,25 +97,10 @@ require('lazy').setup({
       --'gruvbox-community/gruvbox',
       -- 'overcache/NeoSolarized',
     },
-    config = function()
-      setBackground()
-      vim.cmd.colorscheme 'gruvbox'
-
-      -- Hack to avoid first calling togglebg#map on <F5>
-      vim.g.no_plugin_maps = 1
-      vim.fn['togglebg#map']("<F6>")
-      vim.g.no_plugin_maps = nil
-
-      -- react on SigUSR1 to swith between dark and light mode
-      vim.api.nvim_create_autocmd({ "Signal" }, {
-        callback = function()
-          setBackground()
-          vim.cmd.redraw()
-          redrawLine()
-        end
-      })
+    config = function ()
+      require('togglebg-setup').config()
     end
-  },
+    },
   {
     -- Set lualine as statusline
     'nvim-lualine/lualine.nvim',
